+++ R/calc_indicators.R
@@ -22,11 +22,11 @@
 #' @export
 calc_indicators <- function(x, indicators, ...) {
   # check if the requested resource is supported
-  required_resources <- .check_requested_indicator(indicators)
+  req_resources <- .check_requested_indicator(indicators)
   # check if any of the requested resources is already locally available
-  existing_resources <- names(attributes(x)$resources)
+  existing_resources <- names(attributes(x)[["resources"]])
   .check_existing_resources(
-    existing_resources, required_resources,
+    existing_resources, req_resources,
     needed = TRUE
   )
   for (indicator in indicators) x <- .get_single_indicator(x, indicator, ...)
@@ -56,21 +56,21 @@ calc_indicators <- function(x, indicators, ...) {
   # retrieve the selected indicator
   selected_indicator <- available_indicators(indicator)
   # get processing mode
-  processing_mode <- selected_indicator[[indicator]]$processing_mode
+  processing_mode <- selected_indicator[[indicator]][["processing_mode"]]
   # matching the specified arguments to the required arguments
   params <- .check_resource_arguments(selected_indicator, args)
   # append parameters
-  params$verbose <- atts$verbose
-  fun <- selected_indicator[[indicator]]$fun
-  available_resources <- atts$resources
-  required_resources <- selected_indicator[[indicator]]$resources
+  params[["verbose"]] <- atts[["verbose"]]
+  fun <- selected_indicator[[indicator]][["fun"]]
+  avail_resources <- atts[["resources"]]
+  req_resources <- selected_indicator[[indicator]][["resources"]]
 
   processor <- switch(processing_mode,
                       asset = .asset_processor,
                       portfolio = .portfolio_processor,
                       stop(sprintf("Processing mode '%s' is not supported.", processing_mode)))
 
-  results <- processor(x, fun, available_resources, required_resources, params)
+  results <- processor(x, fun, avail_resources, req_resources, params)
   # bind the asset results
   results <- .bind_assets(results)
   # nest the results
@@ -83,27 +83,20 @@ calc_indicators <- function(x, indicators, ...) {
 }
 
 
-.prep <- function(x, available_resources, required_resources) {
-  resources <- purrr::imap(
-    required_resources, function(resource_type, resource_name) {
-      if (resource_type == "raster") {
-        tindex <- read_sf(available_resources[resource_name], quiet = TRUE)
-        resource <- .read_raster_source(x, tindex)
-      } else if (resource_type == "vector") {
-        resource <- .read_vector_source(x, available_resources[[resource_name]])
-      } else {
-        stop(sprintf("Resource type '%s' currently not supported", resource_type))
-      }
-      resource
-    }
-  )
 
-  names(resources) <- names(resources)
-  resources
+.prep_resources <- function(x, avail_resources, req_resources) {
+  if (any(!names(req_resources) %in% names(avail_resources))) {
+    stop("Some required resources are not available.")
+  }
+  purrr::imap(req_resources, function(resource_type, resource_name) {
+    reader <- switch(resource_type,
+                     raster = .read_raster,
+                     vector = .read_vector,
+                     stop(sprintf("Resource type '%s' currently not supported", resource_type)))
+    reader(x, avail_resources[[resource_name]])})
 }
 
-
-.read_vector_source <- function(x, vector_sources) {
+.read_vector <- function(x, vector_sources) {
   vectors <- purrr::map(vector_sources, function(source) {
     tmp <- read_sf(source, wkt_filter = st_as_text(st_as_sfc(st_bbox(x))))
     st_make_valid(tmp)
@@ -112,9 +105,7 @@ calc_indicators <- function(x, indicators, ...) {
   vectors
 }
 
-
-
-.read_raster_source <- function(x, tindex) {
+.read_raster <- function(x, tindex) {
 
   if (st_crs(x) != st_crs(tindex)) {
     x <- st_transform(x, st_crs(tindex))
@@ -157,14 +148,14 @@ calc_indicators <- function(x, indicators, ...) {
 .asset_processor <- function(
     x,
     fun,
-    available_resources,
-    required_resources,
+    avail_resources,
+    req_resources,
     params){
 
   p <- progressr::progressor(steps = nrow(x))
   furrr::future_map(1:nrow(x), function(i) {
     p()
-    resources <- .prep(x[i, ], available_resources, required_resources)
+    resources <- .prep_resources(x[i, ], avail_resources, req_resources)
     result <- .compute(x[i, ], resources, fun, params)
     .check_single_asset(result, i)
   }, .options = furrr::furrr_options(seed = TRUE))
@@ -194,11 +185,11 @@ calc_indicators <- function(x, indicators, ...) {
 .portfolio_processor <- function(
     x,
     fun,
-    available_resources,
-    required_resources,
+    avail_resources,
+    req_resources,
     params ){
 
-  resources <- .prep(x, available_resources, required_resources)
+  resources <- .prep_resources(x, avail_resources, req_resources)
   results <- .compute(x, resources, fun, params)
   if (!inherits(results, "list")) {
     stop("Expected output for processing mode 'portfolio' is a list.")