+++ tests/testthat/test-get_soilgrids.R
@@ -18,7 +18,6 @@ test_that("get_soilgrids works", {
     years = 2000:2020,
     outdir = outdir,
     tmpdir = tmpdir,
-    add_resources = FALSE,
     verbose = FALSE
   )
   # Add testing attribute in order to skip downloads