+++ NEWS.md
@@ -6,6 +6,11 @@
   as an example for how to use the package (#201).
 
   
+## Breaking changes
+- The argument `add_resources` to `init_portfolio()` is now deprecated. This
+  means that `get_resources()` has to be run in every new R session to make 
+  resource available to the respective indicator functions.
+  
 ## Bug Fixes
 
 - `biome` and `ecoregions` now properly handle 0-length tibbles (#196)
@@ -23,10 +28,13 @@
   tiled rasters with and without temporal dimension, single temporal rasters)
 - cropping rasters now uses `snap="out"` by default
 - .read_raster_source now projects assets in case their CRS differs from the raster
+- tile indices for raster resources are now appended to the portfolio attributes
+  as `sf` objects instead of being written to disk (#217)
 - .read_raster_source now applies a precision round-trip of 4 decimal places to
   match rasters with slight changes in their spatial extent (#217)
 - `register_resource()` and `register_indicator()` now issue warnings for 
   resources/indicators with names already registered and overwrites them.
+
 
 
 # mapme.biodiversity 0.4.0